clean up cobs implementation by removing logs and size_t

+++ Common/Libs/Src/UartCobs.cpp
@@ -13,7 +13,7 @@ void UartCobs::write(const uint8_t *buffer, uint16_t length) {
         return;
     }
 
-    size_t encoded_len = cobs_encode(buffer, static_cast<size_t>(length), tx_buf_);
+    uint16_t encoded_len = cobs_encode(buffer, static_cast<uint16_t>(length), tx_buf_);
 
     uart_->write(tx_buf_, static_cast<uint16_t>(encoded_len));
 
@@ -21,12 +21,12 @@ void UartCobs::write(const uint8_t *buffer, uint16_t length) {
     uart_->write(&delimiter, 1);
 }
 
-size_t UartCobs::read(uint8_t *buffer, uint16_t length) {
+uint16_t UartCobs::read(uint8_t *buffer, uint16_t length) {
     if (!uart_ || (!buffer && length != 0)) {
         return 0;
     }
 
-    size_t rx_len = 0;
+    uint16_t rx_len = 0;
     uint8_t byte = 0;
 
     // Read until delimiter OR until rx buffer fills up.
@@ -50,11 +50,11 @@ size_t UartCobs::read(uint8_t *buffer, uint16_t length) {
         return 0;
     }
 
-    size_t decoded_len = 0;
+    uint16_t decoded_len = 0;
     const bool ok = cobs_decode(rx_buf_,
                                 rx_len,
                                 buffer,
-                                static_cast<size_t>(length),
+                                static_cast<uint16_t>(length),
                                 &decoded_len);
 
     return ok ? decoded_len : 0;
